training/requirements.txt
# Training Requirements for CrewAI

# Core ML/AI Libraries
torch>=2.0.0
transformers>=4.30.0  
datasets>=2.12.0
accelerate>=0.20.0
evaluate>=0.4.0

# Training and Evaluation
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
wandb>=0.15.0
tensorboard>=2.13.0

# CrewAI and Related
crewai>=0.28.0
crewai-tools>=0.4.0
langchain>=0.2.0
openai>=1.12.0
litellm>=1.35.0

# Data Processing  
pyyaml>=6.0
jsonlines>=3.1.0
tqdm>=4.65.0
click>=8.1.0

# Development and Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.4.0

# Optional: For advanced training
optuna>=3.2.0         # Hyperparameter optimization
ray>=2.5.0            # Distributed training
mlflow>=2.4.0         # Experiment tracking
